unwrap eth protocol into main fn
